+++ Framework/Audio/Synthesizer/Track.cs
@@ -72,7 +72,7 @@
         /// <param name="pitch">The pitch.</param>
         /// <param name="velocity">The velocity.</param>
         /// <returns>The added note.</returns>
-        public NoteInstance AddNote(ushort beat, ushort length, Note note, Pitch pitch = Pitch.Normal, float velocity = 1f) {
+        public NoteInstance AddNote(float beat, float length, Note note, Pitch pitch = Pitch.Normal, float velocity = 1f) {
             var frequency = new Frequency(note, pitch);
             return this.AddSlideNote(beat, length, frequency, frequency, velocity);
         }
@@ -86,7 +86,7 @@
         /// <param name="endFrequency">The end frequency.</param>
         /// <param name="velocity">The velocity.</param>
         /// <returns>The slide note.</returns>
-        public NoteInstance AddSlideNote(ushort beat, ushort length, Frequency startFrequency, Frequency endFrequency, float velocity) {
+        public NoteInstance AddSlideNote(float beat, float length, Frequency startFrequency, Frequency endFrequency, float velocity) {
             var newNote = new NoteInstance(beat, length, velocity, startFrequency, endFrequency);
             this._notes.Add(newNote);
             return newNote;
